+++ lib/src/ui/int_box.dart
@@ -6,7 +6,7 @@ part of dart_web_toolkit_ui;
 /**
  * A ValueBox that uses {@link IntegerParser} and {@link IntegerRenderer}.
  */
-class IntegerBox extends ValueBox<int> {
+class IntBox extends ValueBox<int> {
 
-  IntegerBox() : super.fromElement(new dart_html.TextInputElement(), new IntRenderer.instance(), new IntParser.instance());
+  IntBox() : super.fromElement(new dart_html.TextInputElement(), new IntRenderer.instance(), new IntParser.instance());
 }